[#16317, #13718] Move common functionality to base iterator class
Summary: Preparatory diff for V2 iterator which introduces base iterator and moves common functionality to this base iterator.

Test Plan:
./yb_build.sh --sj
./build/latest/tests-docdb/docrowwiseiterator-test
Jenkins

Reviewers: timur, rthallam, sergei

Reviewed By: sergei

Subscribers: bogdan

Differential Revision: https://phabricator.dev.yugabyte.com/D24044
